Wiped old project. see description...
We wiped the old project due to it not functioning on the robot and
code was not efficient due to our “get arounds” for robot builder. Now
it has been replaced using our custom robot builder with CANTalons and
wrote code for mecanum. Our mecanum code runs on robot with the
exception of left motors that needed to be inverted. We did finish
inverting the motors, but they ARE NOT TESTED.